skeleton mapped out
the deeper i get into this the more i realize this is wont be a quick project to code. there's a lot i didnt account for and i think ill have to refactor it. Since I have a better understanding of orders, i'm going to put this on the backburner temporarily and enter positions manually until i can figure this app out.